Move resize condition from resizer to interwiki
Removes global-like flag variable that tracked whether to show the
interwiki. Flag started off and switched on when a translation link was
added. Resize attempts would no-op until this was the case.

Moving this logic from the resize function to the main interwiki script
means the resize functions are now logically independent of the
interwiki. When the resize function is called, the iframe resizes. It is
the interwiki's job to decide when to call the function.