Bypass state preservation
Encountered an issue where a browser would not reload the resize iframe
even though the resize details had changes. This was due to state
preservation, where the browser attempts to save a page reload when a
URL hash changes. It is the same reason that changing the hash of a URL
in a normal browser tab does not reload the page. In this case, the
behaviour is incorrect.